datasets/nih_roi/project/roi-consistency/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.1.0
pillow>=10.0.0
numpy>=1.25.0
tqdm>=4.66.0
kagglehub>=0.1.6
